Split cache: Add support for clear-site-data header.
We add a method to extract the resource URL from the
cache key, so that HttpCacheDataRemover can perform
its URL matching to determine which cache entries to
remove. This is in accordance with the spec at
https://w3c.github.io/webappsec-clear-site-data/

In future, it may be desirable to change the spec
so that when there is a clear-site-data request
for a given site, all resources keyed with that
site as the resource URL, the domain, or the
initiator, inclusively, would be removed from the
cache.

Test 1: net_unittests
--gtest_filter=*HttpSplitCacheKeyTest*

Test 2: services_unittests
--gtest_filter=*HttpCacheDataRemoverSplitCacheTest*

Bug: 951812
Change-Id: Id1afd9fff5f766e46d596ce8be340d09e6296bd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1590406
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#666466}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e41d19a7a57341a53284cea3d468fe5028ec70f2