Fix crash that happened in DoesPeerAddressMatchWithFreshAddressList().
The root cause is not clear yet, so the action we are taking right now is to bypass the crash and check at different places for possible causes of crash.

The status of the session is checked and logged.
The status of stale host resolution result is also checked and logged.

Change-Id: I8408ed89dc2187adf20990b86f85ad246d7e4e24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1643496
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Zhongyi Shi <zhongyi@chromium.org>
Commit-Queue: Renjie Tang <renjietang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#666581}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d3465378b5757933913e78637f1038d98a077f7b