Remove old histogram for tracking Android roots
This histogram was introduced when it was not guaranteed
that the Android trust store would be available at the
"AndroidCAStore" KeyStore, which was needed for
determining if a given CA was part of the system base
image. The combination of CTS, plus Chromium moving to
use the //net/data/ssl/root_stores index as a cross-plat
way of determining known root largely obviates the
concerns.

BUG=975152

Change-Id: If9211e6ed22f45042380bcce25b9ccd28057b707
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1663078
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#670099}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fdbdc2acb2a15bf59624979616e38eeb54759277