Add system DNS config refresh capabilities
Used to trigger config update in SystemDnsConfigChangeNotifier on
Android and ChromeOS where the update is triggered external to the
network stack.

Primarily needed for SystemDnsConfigChangeNotifier, but updated the
current soon-to-be-modified code in NetworkChangeNotifierPosix to use
the new unified method on DnsConfigService and removed its special
DnsConfigService overrides in order to simplify and prepare for the
transition to SystemDnsConfigChangeNotifier.

Bug: 971411
Change-Id: I944073d4ff86e43aae40d0e1404a2cfbb1bb902a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1668147
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Eric Orth <ericorth@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#672632}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a97c14ab041e1e3f08082c164eb116d64c1b7cde