Fix URLRequestTest.NetworkDelegateProxyError
The test was supposed to be injecting a MockHostResolver that would
fail all requests into the URLRequestContext, but was never actually
injecting the resolver it created. This could result in the test failing
if there was something listening locally on port 70.

Bug: 496743
Change-Id: Iba8d4899f217c8d10dcfe8d74caf0aba69f8f06d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1680922
Reviewed-by: Eric Orth <ericorth@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#673032}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 39e6de6c20e2205c91733cd1f0111f693ad4933c