Improve performance of GarbageCollectionTriggers, and enable on Win.
CookieMonsterTest.GarbageCollectionTriggers was disabled on Windows,
because it was timing out (Or at least it used to). This CL improves
performance by breaking it into 5 tests, and having CreateMonsterForGC()
not parse 6,600 URLs, which is rather expensive (The split up test that
calls that method is 3 times faster with the optimization).

Bug: 126095
Change-Id: I67f9ad8e2eafa4eb0c80e58cb27ba747c61a9f40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1680600
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#673113}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 477ab63e40b8f4856ccf075839c3eaf57c6de3af