Split cache: Network Isolation key for navigation resources
Adds network isolation key in the ResourceRequest for navigations to be
propagated to net::URLRequest. The key is also updated for redirects.

Updates NavigationRequestInfo to have a network isolation key as a
member instead of top frame origin. Also updated the corresponding
intercepting browser tests in navigation_browsertest.

Added browser tests to verify if the navigation resources (both main
frame and sub frame) are expected to be cached including the redirect use
case.

Also removing the code in url_request_http_job.cc that was using top frame
origin to create NIK if it was not already populated. This is because with
this change we would have NIK being created in browser process for subresources,
navigation resources and dedicated worker resources which were earlier using the
top frame origin.

(Note that sub-resources URL requests do not use this path and have their
network isolation key filled up in the URLLoaderFactory).

Bug: 984130
Change-Id: I4c247b91441516018605723b73b4908536c19ab1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1665549
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Commit-Queue: Shivani Sharma <shivanisha@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#677937}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b5366c62c02fd6edfe83703e57e4fcd14d38f64