Send minimal HTTP headers for DNS-over-HTTPS requests.
Use a string constant for the User-Agent header and a wildcard for the
Accept-Language header. This reduces the ability of providers to group
traffic from a given user across connections but allows providers to
trace the queries back to Chrome in debugging efforts.

Bug: 878582
Change-Id: I9c51b7400640f7df7c0d51cb6290d6a37f564f4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1689398
Commit-Queue: Katharine Daly <dalyk@google.com>
Reviewed-by: Eric Orth <ericorth@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#678521}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5afc14e66e450c0ffc2eac787d4cc8569d224783