Records the count of requests to UMA
The peak queueing delay within a measurement period should increase if
the count of in-flight requests increases. To verify this, Chromium
records the count of in-flight requests causing the peak queueing delay
to UMA.

Adds the code to record the count of in-flight requests keyed by the
queueing delay levels.

Change-Id: I98225cf0e35bcbc6ff6ef7ca9052013e14460d6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1707574
Commit-Queue: Jianfeng Wang <jfwang@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#679253}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dbae90ffd3365ec92e1136d6813a522ec83df88a