Remove Secure, SameSite, HttpOnly checks from CanonicalCookie::Create()
Create() does not need to check IsSetPermittedInContext because
it is already checked when setting a cookie. It also does not have
to check for a secure scheme when creating a Secure cookie, because
that is checked when setting a cookie. Create() should only
need to check that the cookie it creates is canonical.

Bug: 989098
Change-Id: Ic5be164cfb60825c7775c3bc58986a835afa499e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1726789
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#683002}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f5732c8ad2485883822d3a3b2bbb7b7c71390e59