Pass top frame origin to cookie settings
In order to correctly apply content settings with secondary patterns,
CookieSettings needs to know the TopFrameOrigin of a request.
Currently it only has access to SiteForCookies, which can be an empty
url for cross-origin iframes.
This CL passes TopFrameOrigin from UrlRequests to CookieSettings and
uses this value to look up ContentSettings. SiteForOrigin is still
used to determine whether the cookie is used in a third-party context.

As TopFrameOrigin is not yet always available, this new behavior is
behind the ImprovedCookieControls flag.

Bug: 988398
Change-Id: Ibb6180b68ae9f4780046be6fb2a9dcbb39f9ccf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1698419
Commit-Queue: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Roger Tawa <rogerta@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#686759}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b21dff94e34c617b52a335ecf531b8010c9a3fa4