Introduce //net RESTRICTED_PREFETCH load flags
Presently, prefetch responses in the Http Cache can be reused by any
matching request. This CL introduces a RESTRICTED_PREFETCH load flag
that requests should set if their reuse must be limited to requests
with the CAN_USE_RESTRICTED_PREFETCH flag.

This is necessary to support cross-origin prefetches, whose reuse must
be limited to top-level navigations. This CL is meant to be a precursor
to the cross-origin prefetch implementation work that uses the flags
introduced in the CL.

This CL also includes tests for the flags, ensuring that only requests
adorned with the CAN_USE_RESTRICTED_PREFETCH load flag can reuse
restricted prefetches. Furthermore, it also tests that if request not
adorned with CAN_USE_RESTRICTED_PREFETCH attempts to reuse a restricted
prefetch, the cache will be skipped, and the resource will be retrieved
from the network. The new cache entry will not be marked as restricted.

Bug: 939317
Change-Id: I30a1fc21cc91175fe937450db27109c154861f70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1735159
Commit-Queue: Dominic Farolino <dom@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#687582}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4ecaf0a3b4426a857cefc3f9fcb450a6e917036a