HttpServerProperties: Add NetworkIsolationKey as ServerInfo index.
ServerInfoMaps now use NetworkIsolationKeys in addition to
SchemeHostPorts as indices. These new indices are stored on disk. This
CL also introduces a feature to enable respecting these indices when
creating the keys.

The on-disk format of the ServerInfoMap changes from:
[{<server1>: {<values1>}},{<server2>: {<values2>}}]
to
[{<values1>},{<values2>}]

And <values> dictionaries now contain both the server HostPortPairs,
and the NetworkIsolationKeys. As a result of this change, ServerInfo
data from old prefs files will be lost, though other entries will
be retained.

This CL does not add an API to actually set NetworkIsolationKey when
using an HttpServerProperties object, that will come in another CL.

Bug: 969890
Change-Id: I8b115f5e704358c3178052e225087de2ef5446a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1745820
Reviewed-by: Zhongyi Shi <zhongyi@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#688576}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1be93d283a98cf59e900aa993a8b7cf5c7f124e9