Replace some uses of 0.0.0.0 in tests which expect failure.
Connecting to 0.0.0.0 varies between platforms. Tests that intend to have connections fail should not use it, as it is interpreted as the ANY address on Linux/Mac and connections to it may succeed leading to test flakiness.

0.0.0.1 is a more reliable address for testing. It is in the same reserved IPv4 block, but does not have special meaning like 0.0.0.0.

0.0.0.0 has a consistent meaning when listening, so fine to use for server sockets.

Change-Id: If484cb42023b2d85e2517f959fb625accaf87161
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1764941
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Eric Roman <eroman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#689640}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 862ae4b2a1a00252baee0ac117e8c24a9cd6faee