Convert DCHECK to CHECK to validate cause of HttpCacheTransaction bug
DoneHeadersAddToEntryComplete assumes that the previous result is OK.
However, it may be ERR_CACHE_RACE in some situations, resulting in a
crash. Converting to a CHECK would verify that this is the cause of
some crashes observed in the wild.

Bug: 959194
Change-Id: Ib989e85b5334667875e8669f442f32c274b0d23c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1773163
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#690762}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 64bc8686d3398ed8c316143ad8d9d6c94f5cccf8