HttpServerProperties: Add NetworkIsolationKey to RequiresHTTP11 methods.
This CL adds NetworkIsolationKey as an argument to the
RequiresHTTP11 methods in HttpServerProperties, and makes consumers
pass in the correct key. It also moves tracking of this bit over to the
ServerInfo struct, to simplify things a bit, which requires the
RequiresHTTP11 take a SchemeHostPort instead of a HostPortPair.

Bug: 969890
Change-Id: Ieb7a8ae234eae2d459895b63101be4d1c00d82fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1761553
Reviewed-by: Zhongyi Shi <zhongyi@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#691444}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f2ee07ced0a6156b2aa8e9656e842566b62dfca2