Call RequestHeadersCallback on internal redirects
On HSTS redirects, a new URLRequestRedirectJob is made in such a way
that calls the `responseReceivedExtraInfo` DevTools protocol event,
but not the `requestWillBeSentExtraInfo` event called from the request
headers callback. This calls the callback with fake headers to for the
matching fake headers sent out to the other event.

Bug: 993019
Change-Id: Ibd1645f4ba1458557095e992d910e3aedfdbbad9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1758341
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Aaron Tagliaboschi <aarontag@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#693330}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 22e1b3a12e4080e49ae71424b7ff225de2e4a5f9