Split out CookieEffectiveSameSite enum from CookieSameSite
This makes CookieSameSite into two separate enums, CookieSameSite
(representing the SameSite attribute value of the parsed cookie), and
CookieEffectiveSameSite (representing the SameSite mode to be applied
to inclusion/exclusion of that cookie).

Previously they were the same enum, which resulted in overlapping sets
of enum values that were possibly valid for one but not the other.
This change removes the need to check whether a value is valid, because
only valid semantically values are now included in the respective enums.

Bug: None
Change-Id: I467e7ac0551511e1afea74828886f40a57065d2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1808019
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#697462}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: da41c3d35f495f18a30b2a9ac091d21996936bff