Include CookieAccessSemantics in effective SameSite calculation
This change hooks up the CookieMonster's CookieAccessDelegate to
CanonicalCookie::IncludeForRequestURL and IsSetPermittedInContext, by
passing them the CookieAccessSemantics for that particular cookie that
is returned by the delegate. The access semantics is fed into
CanonicalCookie::GetEffectiveSameSite, where it is taken into account
when computing the effective SameSite mode of a cookie.

For now, the default value of the parameter is CookieAccessSemantics::
UNKNOWN, so other callers do not yet pass the right access semantics
to IncludeForRequestURL and IsSetPermittedInContext. They will be
updated in subsequent changes.

Bug: 978172
Change-Id: Icf0dc68738c3651c8e0fc648e597658485dc799c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1845233
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#703490}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 70f997f771f999d02882d3857954c5927d3bb405