Block TE, Keep-Alive, Cookie2, Transfer-Encoding headers.
None of those headers are set often enough in the field to make blocking
them problematic. Also remove logging of those headers, User-Agent,
and Content-Transfer-Encoding. The latter two are now allowed to be set
by the latest version of the fetch spec.

Also remove Content-Transfer-Encoding from the network stack's
kForbiddenHeaderFields (for the aforementioned reason), but leave in
User-Agent, since there's an issue around ServiceWorkers allowing it to
be set on cross-site requests. Also add "DNT" to the list, to match
the fetch spec.

"DNT" is not added to the block list, since it's currently set in the
renderer.

Bug: 974911
Change-Id: Ie0f75ccd6f6c225031157db34c6e9a2d3be9972f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1827383
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Bill Budge <bbudge@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#703891}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3611052c055897e5ebbc5b73ea295092e0c20141